Move Server and tunnel setup before cart creation.
